# Refactor for performance