--- a/src/main/java/com/wastedbankspace/WastedBankSpaceConfig.java
+++ b/src/main/java/com/wastedbankspace/WastedBankSpaceConfig.java
@@ -28,15 +28,22 @@
 
 package com.wastedbankspace;
 
+import com.google.common.collect.Sets;
 import com.wastedbankspace.ui.overlay.OverlayImage;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+import lombok.Getter;
 import net.runelite.client.config.Config;
 import net.runelite.client.config.ConfigGroup;
 import net.runelite.client.config.ConfigItem;
 import net.runelite.client.config.ConfigSection;
 
-@ConfigGroup("Wasted Bank Space")
+@ConfigGroup(WastedBankSpaceConfig.GROUP)
 public interface WastedBankSpaceConfig extends Config
 {
+	String GROUP = "Wasted Bank Space";
 	/* Storage Location Keys */
 	String CLUE_ITEM_CHECK_KEY = "clueItemCheck";
 	String FANCY_DRESS_BOX_KEY = "poHFancyDressBox";
@@ -64,33 +71,57 @@ public interface WastedBankSpaceConfig extends Config
 	String FILTER_ENABLED_CHECK_KEY = "filterEnabledCheck";
 	String BIS_FILTER_ENABLED_CHECK_KEY = "bisFilterEnabledCheck";
 
+	static HashSet<String> getStorageLocationKeys() {
+		return Sets.newHashSet(
+			CLUE_ITEM_CHECK_KEY,
+			FANCY_DRESS_BOX_KEY,
+			MAGIC_WARDROBE_KEY,
+			SEED_CHECK_KEY,
+			TACKLE_BOX_CHECK_KEY,
+			STEEL_KEY_RING_CHECK_KEY,
+			TOOL_LEP_CHECK_KEY,
+			MASTER_SCROLL_BOOK_CHECK_KEY,
+			FOSSIL_STORAGE_CHECK_KEY,
+			ELNOCK_INQUISITOR_CHECK_KEY,
+			FLAMTAER_BAG_CHECK_KEY,
+			NIGHTMARE_ZONE_CHECK_KEY,
+			TOY_BOX_CHECK_KEY,
+			SPICE_RACK_CHECK_KEY,
+			ARMOUR_CASE_CHECK_KEY,
+			FORESTRY_KIT_CHECK_KEY,
+			MYSTERIOUS_STRANGER_CHECK_KEY,
+			PET_HOUSE_SPACE_CHECK_KEY,
+			HOUSE_BOOKCASE_CHECK_KEY,
+			CAPE_RACK_CHECK_KEY,
+			HUNTSMANS_KIT_SPACE_CHECK_KEY);
+	}
 
 	@ConfigSection(
-			name = "General Config",
-			description = "Genera plugin config flags",
-			position = 0
+		name = "General Config",
+		description = "Genera plugin config flags",
+		position = 0
 	)
 	String generalConfig = "generalConfig";
 
 	@ConfigSection(
-			name = "Non Flagged Items",
-			description = "Items blacklisted for being flagged",
-			position = 1
+		name = "Non Flagged Items",
+		description = "Items blacklisted for being flagged",
+		position = 1
 	)
 	String nonFlaggedItems = "nonFlaggedItems";
 
 	@ConfigSection(
-			name = "Bank Interface Config",
-			description = "Configure bank interface changes",
-			position = 2
+		name = "Bank Interface Config",
+		description = "Configure bank interface changes",
+		position = 2
 	)
 	String bankInterfaceConfig = "bankInterfaceConfig";
 
 	@ConfigItem(keyName = FILTER_ENABLED_CHECK_KEY,
-			name = "Enable Filtering",
-			description = "Enable Item Filters and Blacklist",
-			position = 0,
-			section = nonFlaggedItems
+		name = "Enable Filtering",
+		description = "Enable Item Filters and Blacklist",
+		position = 0,
+		section = nonFlaggedItems
 	)
 	default boolean filterEnabledCheck()
 	{
@@ -98,22 +129,22 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = BIS_FILTER_ENABLED_CHECK_KEY,
-			name = "Never Filter BIS",
-			description = "Never Filter Best in Slot Items",
-			position = 1,
-			section = nonFlaggedItems
+		name = "Never Filter BIS",
+		description = "Never Filter Best in Slot Items",
+		position = 1,
+		section = nonFlaggedItems
 	)
-	default boolean bisfilterEnabledCheck()
+	default boolean bisFilterEnabledCheck()
 	{
 		return true;
 	}
 
 	@ConfigItem(
-			keyName = "nonFlaggedItems",
-			name = "",
-			description = "Configures items to not be flagged as wasted bank space. Format: item, item",
-			section = nonFlaggedItems,
-			hidden = true
+		keyName = "nonFlaggedItems",
+		name = "",
+		description = "Configures items to not be flagged as wasted bank space. Format: item, item",
+		section = nonFlaggedItems,
+		hidden = true
 	)
 	default String nonFlaggedItems()
 	{
@@ -121,17 +152,17 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(
-			keyName = "nonFlaggedItems",
-			name = "",
-			description = "Setter for storing notes data"
+		keyName = "nonFlaggedItems",
+		name = "",
+		description = "Setter for storing notes data"
 	)
 	void nonFlaggedItems(String str);
 
 	@ConfigItem(keyName = CLUE_ITEM_CHECK_KEY,
-			name = "PoH Clue Item Storage",
-			description = "Treasure Chest Storage",
-			position = 1,
-			section = generalConfig
+		name = "PoH Clue Item Storage",
+		description = "Treasure Chest Storage",
+		position = 1,
+		section = generalConfig
 	)
 	default boolean treasureChestStorageCheck()
 	{
@@ -139,10 +170,10 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = FANCY_DRESS_BOX_KEY,
-			name = "PoH Fancy Dress Box",
-			description = "Random Event Items",
-			position = 2,
-			section = generalConfig
+		name = "PoH Fancy Dress Box",
+		description = "Random Event Items",
+		position = 2,
+		section = generalConfig
 	)
 	default boolean fancyDressBoxStorageCheck()
 	{
@@ -150,10 +181,10 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = MAGIC_WARDROBE_KEY,
-			name = "PoH Magic Wardrobe Storage",
-			description = "Magic armour and robe sets",
-			position = 3,
-			section = generalConfig
+		name = "PoH Magic Wardrobe Storage",
+		description = "Magic armour and robe sets",
+		position = 3,
+		section = generalConfig
 	)
 	default boolean magicWardrobeStorageCheck()
 	{
@@ -161,10 +192,10 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = SEED_CHECK_KEY,
-			name = "Seed Box/Vault",
-			description = "Seeds can be stored in a seed box, saplings and seeds in the vault",
-			position = 4,
-			section = generalConfig
+		name = "Seed Box/Vault",
+		description = "Seeds can be stored in a seed box, saplings and seeds in the vault",
+		position = 4,
+		section = generalConfig
 	)
 	default boolean seedVaultStorageCheck()
 	{
@@ -172,10 +203,10 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = TACKLE_BOX_CHECK_KEY,
-			name = "Tackle Box Items",
-			description = "Check for items that could be in your tackle box",
-			position = 5,
-			section = generalConfig
+		name = "Tackle Box Items",
+		description = "Check for items that could be in your tackle box",
+		position = 5,
+		section = generalConfig
 	)
 	default boolean tackleBoxStorageCheck()
 	{
@@ -183,10 +214,10 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = STEEL_KEY_RING_CHECK_KEY,
-			name = "Steel Key Ring",
-			description = "Stores some keys - One small favor reward",
-			position = 6,
-			section = generalConfig
+		name = "Steel Key Ring",
+		description = "Stores some keys - One small favor reward",
+		position = 6,
+		section = generalConfig
 	)
 	default boolean steelKeyRingStorageCheck()
 	{
@@ -194,10 +225,10 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = TOOL_LEP_CHECK_KEY,
-			name = "Tool Leprechaun",
-			description = "Stores farming items, note low limits on many items",
-			position = 7,
-			section = generalConfig
+		name = "Tool Leprechaun",
+		description = "Stores farming items, note low limits on many items",
+		position = 7,
+		section = generalConfig
 	)
 	default boolean toolLeprechaunStorageCheck()
 	{
@@ -205,10 +236,10 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = MASTER_SCROLL_BOOK_CHECK_KEY,
-			name = "Master Scroll Book",
-			description = "Stores teleport scrolls",
-			position = 8,
-			section = generalConfig
+		name = "Master Scroll Book",
+		description = "Stores teleport scrolls",
+		position = 8,
+		section = generalConfig
 	)
 	default boolean masterScrollBookStorageCheck()
 	{
@@ -216,10 +247,10 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = FOSSIL_STORAGE_CHECK_KEY,
-			name = "Fossil Storage",
-			description = "Storage for fossils on Fossil Island and in Varrok Museum",
-			position = 9,
-			section = generalConfig
+		name = "Fossil Storage",
+		description = "Storage for fossils on Fossil Island and in Varrock Museum",
+		position = 9,
+		section = generalConfig
 	)
 	default boolean fossilStorageStrorageCheck()
 	{
@@ -227,10 +258,10 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = ELNOCK_INQUISITOR_CHECK_KEY,
-			name = "Puro Puro Storage",
-			description = "Elnock Inquisitor will hold some items in Puro Puro",
-			position = 10,
-			section = generalConfig
+		name = "Puro Puro Storage",
+		description = "Elnock Inquisitor will hold some items in Puro Puro",
+		position = 10,
+		section = generalConfig
 	)
 	default boolean elnockInquisitorStorageCheck()
 	{
@@ -238,10 +269,10 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = FLAMTAER_BAG_CHECK_KEY,
-			name = "Flamtaer Bag Storage",
-			description = "Storage for Shades of Morton (minigame) items",
-			position = 11,
-			section = generalConfig
+		name = "Flamtaer Bag Storage",
+		description = "Storage for Shades of Morton (minigame) items",
+		position = 11,
+		section = generalConfig
 	)
 	default boolean flamtaerBagStorageCheck()
 	{
@@ -249,10 +280,10 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = NIGHTMARE_ZONE_CHECK_KEY,
-			name = "Nightmare Zone Potion Storage",
-			description = "Potion's which store in the barrels at the nmz (minigame)",
-			position = 12,
-			section = generalConfig
+		name = "Nightmare Zone Potion Storage",
+		description = "Potion's which store in the barrels at the nmz (minigame)",
+		position = 12,
+		section = generalConfig
 	)
 	default boolean nightmareZoneStorageCheck()
 	{
@@ -260,32 +291,32 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = TOY_BOX_CHECK_KEY,
-			name = "PoH Toy Box Storage",
-			description = "Toy and holiday item storage",
-			position = 13,
-			section = generalConfig
+		name = "PoH Toy Box Storage",
+		description = "Toy and holiday item storage",
+		position = 13,
+		section = generalConfig
 	)
-	default	boolean toyBoxStorageCheck()
+	default boolean toyBoxStorageCheck()
 	{
 		return true;
 	}
 
 	@ConfigItem(keyName = SPICE_RACK_CHECK_KEY,
-			name = "Spice rack storage",
-			description = "Storage for spicy stew spices in the PoH kitchen",
-			position = 14,
-			section = generalConfig
+		name = "Spice rack storage",
+		description = "Storage for spicy stew spices in the PoH kitchen",
+		position = 14,
+		section = generalConfig
 	)
 	default boolean spiceRackStorageCheck()
 	{
-    	return true;
+		return true;
 	}
 
 	@ConfigItem(keyName = FORESTRY_KIT_CHECK_KEY,
-			name = "Forestry kit storage",
-			description = "Storage for Forestry-related items",
-			position = 15,
-			section = generalConfig
+		name = "Forestry kit storage",
+		description = "Storage for Forestry-related items",
+		position = 15,
+		section = generalConfig
 	)
 	default boolean forestryKitStorageCheck()
 	{
@@ -293,39 +324,43 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = ARMOUR_CASE_CHECK_KEY,
-			name = "PoH Armour Case storage",
-			description = "Armour and outfit storage",
-			position = 16,
-			section = generalConfig
+		name = "PoH Armour Case storage",
+		description = "Armour and outfit storage",
+		position = 16,
+		section = generalConfig
 	)
 	default boolean armourCaseStorageCheck()
 	{
-		return false; //Only for people to turn on if they want it on, nobody wants BIS gear with little triangles
+		return false; // Only for people to turn on if they want it on, nobody wants BIS gear with little triangles
 	}
 
 	@ConfigItem(keyName = MYSTERIOUS_STRANGER_CHECK_KEY,
-			name = "Mysterious Stranger",
-			description = "Storage for Hallowed Sepulchre items",
-			position = 17,
-			section = generalConfig
+		name = "Mysterious Stranger",
+		description = "Storage for Hallowed Sepulchre items",
+		position = 17,
+		section = generalConfig
 	)
-	default boolean mysteriousStrangerStorageCheck() { return true; }
+	default boolean mysteriousStrangerStorageCheck()
+	{
+		return true;
+	}
 
 	@ConfigItem(keyName = PET_HOUSE_SPACE_CHECK_KEY,
-			name = "Pet House Space",
-			description = "Pet House Space inside the POH",
-			position = 18,
-			section = generalConfig
+		name = "Pet House Space",
+		description = "Pet House Space inside the POH",
+		position = 18,
+		section = generalConfig
 	)
-	default boolean petHouseStorageCheck() {
-		return false; //Disabled by Default. Most people don't lose pets in their bank alike BIS gear.
+	default boolean petHouseStorageCheck()
+	{
+		return false; // Disabled by Default. Most people don't lose pets in their bank alike BIS gear.
 	}
 
 	@ConfigItem(keyName = HOUSE_BOOKCASE_CHECK_KEY,
-			name = "PoH Bookcase",
-			description = "Bookcase inside the PoH",
-			position = 19,
-			section = generalConfig
+		name = "PoH Bookcase",
+		description = "Bookcase inside the PoH",
+		position = 19,
+		section = generalConfig
 	)
 	default boolean bookcaseStorageCheck()
 	{
@@ -333,33 +368,33 @@ public interface WastedBankSpaceConfig extends Config
 	}
 
 	@ConfigItem(keyName = CAPE_RACK_CHECK_KEY,
-			name = "Cape rack storage",
-			description = "Storage for capes in the costume room",
-			position = 20,
-			section = generalConfig
+		name = "Cape rack storage",
+		description = "Storage for capes in the costume room",
+		position = 20,
+		section = generalConfig
 	)
-    default boolean capeRackStorageCheck()
+	default boolean capeRackStorageCheck()
 	{
 		return true;
-  	}
+	}
 
 	@ConfigItem(keyName = HUNTSMANS_KIT_SPACE_CHECK_KEY,
-			name = "Huntsman's Kit",
-			description = "Check for items that could go in your Huntsman's kit",
-			position = 21,
-			section = generalConfig
+		name = "Huntsman's Kit",
+		description = "Check for items that could go in your Huntsman's kit",
+		position = 21,
+		section = generalConfig
 	)
-    default boolean huntsmansKitStorageCheck()
+	default boolean huntsmansKitStorageCheck()
 	{
 		return true;
 	}
 
 	@ConfigItem(
-			keyName = "overlayImage",
-			name = "Overlay Image",
-			description = "Configures images to overlay on items wasting space",
-			position = 0,
-			section = bankInterfaceConfig
+		keyName = "overlayImage",
+		name = "Overlay Image",
+		description = "Configures images to overlay on items wasting space",
+		position = 0,
+		section = bankInterfaceConfig
 	)
 	default OverlayImage overlayImage()
 	{