--- a/src/main/java/com/wastedbankspace/model/StorageLocations.java
+++ b/src/main/java/com/wastedbankspace/model/StorageLocations.java
@@ -135,7 +135,7 @@ public class StorageLocations
 	}
 	public static String getStorableItemName(Integer id)
 	{
-		return getStorableItemName(itemIdMap.get(id));
+		return itemNameMap.get(id);
 	}
 
 	public static String getStorableItemName(StorableItem item)