… src/main/java/com/wastedbankspace/model/locations/ArmourCase.java
… src/main/java/com/wastedbankspace/ui/WastedBankSpacePanel.java
… src/main/java/com/wastedbankspace/ui/overlay/OverlayImage.java
… src/main/java/com/wastedbankspace/ui/overlay/StorageItemOverlay.java