Refactored heavy function calls to significantly improve performance, especially on high refresh rate monitors when runelite is set to render above 60fps.

Gains realized:

- 45fps before -> 105fps after
- 105fps with plugin disabled


More context and screenshots of improvements can be found in #84