@mcgeer 2025-09-09
@swirle13 Will soon TM. TL;DR, next Monday.

I got like half way through the first weekend, then got told I was moving forward with some interviews and no-life'd leet code.

I am away this weekend I hate to push this a month. I can potentially do next Monday? If that works for you. I have some comments locally, but things I was just gonna address (Dead code and such) I havent spotted major things as of yet.