@mcgeer 2025-09-20
@swirle13  https://github.com/runelite/plugin-hub/pull/9100