@swirle13 2025-09-20
Hey, thank you so much for the time spent testing and the fixes! I didn't have much spare time these past two weeks, so I appreciate all your work to get this integrated and working properly. Thank you!